Add Purple Heart, Edit expression names, Added model preview, model switch, model scale parameter
