+++ README.md
@@ -1,11 +1,13 @@
 <div align="center" style="padding-top: 10px">
-    <img src="./docs/nep.svg" alt="koe" width="50px" height="50px">
-    <img src="./docs/nep_logo.svg" alt="live2d-py" height="40px">
+    <img src="./docs/nep.svg" alt="nep" width="50px" height="50px">
+    <img src="./docs/nep_logo.svg" alt="my_little_neptune" height="40px">
 </div>
 <p align="center" style="font-family: 'Roboto', sans-serif; font-size: 1em; color: #555;">
     <br>
-    <img title="Python Version" src="https://img.shields.io/badge/python-3.10+-blue" alt="Python Version" style="margin: 0 10px;">
-    <img title="App Version" src="https://img.shields.io/badge/release-0.1.1-purple" alt="App Version" style="margin: 0 10px;">
+    <img title="Python Version" src="https://img.shields.io/badge/Python-3.1.2-blue" alt="Python Version" style="margin: 0 10px;">
+    <img title="PySide6 Version" src="https://img.shields.io/badge/PySide6-6.8.1.1-green" alt="PySide6 Version" style="margin: 0 10px;">
+    <img title="live2d-py Version" src="https://img.shields.io/badge/live2d-0.3.4-orange" alt="live2d-py Version Version" style="margin: 0 10px;">
+    <img title="App Version" src="https://img.shields.io/badge/version-0.1.2-purple" alt="App Version" style="margin: 0 10px;">
 </p>
 
 ## The assistant application on your desktop, which pleases you with its appearance every day:)
@@ -22,4 +24,23 @@
 
 ## Usage:
 Run:
-`python package/neptune_main_pyside6.py`
+`python package/neptune_main.py`
+
+### Models Select:
+Edit `self.models_switch = 0` parameter in package/neptune_main.py file, to the model I'd:
+* Neptune = 0
+* Purple Heart = 1
+### Models Scale:
+Edit `self.models_scale = 1` parameter in package/neptune_main.py file,  to scale model
+
+## Models Available:
+<div align="center" style="padding-top: 10px">
+    <img src="./docs/model_preview/neptune.svg" alt="neptune" width="300px" height="300px">
+    <img src="./docs/model_preview/purple_heart.svg" alt="purple_heart" width="300px" height="300px">
+</div>
+
+## Important to read:
+#### 1. The application is at an early stage of development and may have bugs.
+#### 2. The animation of the models needs to be improved
+#### 3. In the next stages of development, it is planned to work on the GUI, AI assistant base on GPT model and compile the application into an .exe file
+### Thanks for your attention!
