--- Resources/v3/Neptune/expressions/Tired.json
@@ -1,12 +1,0 @@
-{
-  "Type": "Live2D Expression",
-  "FadeInTime": 0.5,
-  "FadeOutTime": 0.5,
-  "Parameters": [
-    {
-      "Id": "Param19",
-      "Value": 1,
-      "Blend": "Add"
-    }
-  ]
-}
