+++ Resources/v3/PurpleHeart/expressions/CloseEyes.json
@@ -1,0 +1,12 @@
+{
+  "Type": "Live2D Expression",
+  "FadeInTime": 0.5,
+  "FadeOutTime": 0.5,
+  "Parameters": [
+    {
+      "Id": "Param18",
+      "Value": 1,
+      "Blend": "Add"
+    }
+  ]
+}
