+++ Resources/v3/PurpleHeart/expressions/Funny.json
@@ -1,0 +1,12 @@
+{
+  "Type": "Live2D Expression",
+  "FadeInTime": 0.5,
+  "FadeOutTime": 0.5,
+  "Parameters": [
+    {
+      "Id": "Param23",
+      "Value": 1,
+      "Blend": "Add"
+    }
+  ]
+}
