+++ Resources/v3/PurpleHeart/expressions/Normal.json
@@ -1,0 +1,62 @@
+{
+  "Type": "Live2D Expression",
+  "FadeInTime": 0.5,
+  "FadeOutTime": 0.5,
+  "Parameters": [
+    {
+      "Id": "PARAM_RAGE",
+      "Value": 0,
+      "Blend": "Add"
+    },
+    {
+      "Id": "Param8",
+      "Value": 0,
+      "Blend": "Add"
+    },
+    {
+      "Id": "Param9",
+      "Value": 0,
+      "Blend": "Add"
+    },
+    {
+      "Id": "Param12",
+      "Value": 0,
+      "Blend": "Add"
+    },
+    {
+      "Id": "Param14",
+      "Value": 0,
+      "Blend": "Add"
+    },
+    {
+      "Id": "Param18",
+      "Value": 0,
+      "Blend": "Add"
+    },
+    {
+      "Id": "Param19",
+      "Value": 0,
+      "Blend": "Add"
+    },
+    {
+      "Id": "Param20",
+      "Value": 0,
+      "Blend": "Add"
+    },
+    {
+      "Id": "Param21",
+      "Value": 0,
+      "Blend": "Add"
+    },
+    {
+      "Id": "Param22",
+      "Value": 0,
+      "Blend": "Add"
+    },
+    {
+      "Id": "Param23",
+      "Value": 0,
+      "Blend": "Add"
+    }
+  ]
+}
