+++ Resources/v3/PurpleHeart/expressions/Surprised.json
@@ -1,0 +1,12 @@
+{
+  "Type": "Live2D Expression",
+  "FadeInTime": 0.5,
+  "FadeOutTime": 0.5,
+  "Parameters": [
+    {
+      "Id": "Param22",
+      "Value": 1,
+      "Blend": "Add"
+    }
+  ]
+}
